+++ components/donate/donorprofile/DonorProfile.jsx
@@ -1,5 +1,5 @@
-import DonateNav from '../donate/DonateNav'
-import DonateFooter from '../donate/DonateFooter'
+import DonateNav from '../DonateNav'
+import DonateFooter from '../DonateFooter'
 import DonorProfileBody from './DonorProfileBody'
 
 const DonorProfile = () => {
